+++ app/Basket/Synchronisation/MapApplicationHelper.php
@@ -52,6 +52,7 @@ class MapApplicationHelper
         $application->ext_metadata = json_encode($applicationEntity->getMetadata());
         $application->ext_user = $applicationEntity->getUser();
         $application->ext_is_regulated = is_null($applicationEntity->getIsRegulated()) ? $application->ext_is_regulated : $applicationEntity->getIsRegulated();
+        $application->ext_merchant_liable_at = is_null($applicationEntity->getMerchantLiableAt()) ? $applicationEntity->getMerchantLiableAt() : Carbon::parse($applicationEntity->getMerchantLiableAt());
     }
 
     /**