+++ database/migrations/2017_06_09_120000_add_ext_feature_merchant_liable_to_installations_table.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+/**
+ * Add Ext Feature Merchant Liable To Installations Table
+ *
+ * @author EB
+ */
+class AddExtFeatureMerchantLiableToInstallationsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('installations', function (Blueprint $table) {
+            $table->boolean('ext_feature_merchant_liable')->after('ext_default_product')->default(false);
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('installations', function (Blueprint $table) {
+            $table->dropColumn('ext_feature_merchant_liable');
+        });
+    }
+}